requirements.txt
Django==3.1.12

django-admin-honeypot==1.1.0
django-celery-beat==2.0.0
django-csp==3.6
django-extensions==3.0.1
django-nested-admin==3.3.2
django-otp==0.9.3
django-referrer-policy==1.0
django-tabbed-admin==1.0.4
django-import-export==2.5.0
django-debug-toolbar==3.2.1
django-js-reverse==0.9.1

phonenumberslite==8.12.21
django-phonenumber-field==5.0.0
ovh==0.5.0

black==20.8b1
celery[redis]==4.4.6
coverage==5.1
entrypoints==0.3
factory-boy==2.12.0
flake8==3.8.3
freezegun==0.3.15
gunicorn==20.0.4
ipython==7.16.1
mock==4.0.2
Pillow==8.2.0
pre-commit==2.11.1
psycopg2-binary==2.8.5
ptpython==3.0.2
pudb==2019.2
PyJWT==1.7.1
python-dotenv==0.13.0
pytz==2020.1
qrcode==6.1
requests==2.24.0
selenium==3.141.0
sentry-sdk==1.0.0
sqlparse==0.3.1
whitenoise==5.1.0

git+git://github.com/betagouv/django-magicauth@0.7
